Quick fix for deadlock in MetricDataCache.  The synchronization in this class still needs work.
Change getKeysWithExpiryCheck to just getKeys since no expiration occurs in this cache.  This was
causing a lock on the Cache for getKeys, then a lock for each Element as the expiration was checked.
The bulkAdd locks the Element, then the entire cache if a remove() operation is needed, which was
causing the deadlock.
[HHQ-1288]