[HQ-1158] Fixed the problem with percentage conditions. Should use the min of the calculated percentages of violators and 100, not the max of the two.
Changed percentage conditions to be based on the total group size.

Fixed time window so that the time window may be shifted to the next start time of a heart beat as well as a measurement event. This is to minimize the chance of the time window shifting too rapidly so that measurement events are always considered older than the time window.

Fixed the amount of time we wait for events to be populated in the initial time window before advancing forward so that we use exact times, not voodooed down times. The voodooing down could potentially reduce the amount of time we wait by 1 minute.

Updated gtrigger debug logging to include the name of the containing galert as well as whether or not this trigger is assigned a NORMAL or RECOVERY partition.