[HHQ-1603] Based on performance results indicating that the single txn per trigger event processing model doesn't scale, rolled back this txn processing model applied for HQ-1115. With the single txn per trigger event processing we were seeing a high load average and JMS queue backup at a large customer installation.
The only difference with the rolled back version is that now we are *not* catching Throwable around RegisteredDispatcherEJBImpl.tryFlushState() in RegisteredDispatcherEJBImpl.flushStateForVisitedMCTriggers() since the single txn should be rolled back if an exception occurs when flushing the trigger state. In this case we wouldn't want to continue processing the events.