Refactored duration trigger into a simple state machine with explicit states.
We're now tracking all TriggerFiredEvents and TriggerNotFiredEvents so we don't lose resolution on the tracked events when old events age out.

Instead of buffering the tracked events 30 seconds (setting expiration time 30 seconds later), we are estimating the buffer time based off the collection interval for the trackable events sent to the duration trigger.

Added hooks to make trigger unit testable.