[HHQ-3110] - Alert condition value is trimmed after first decimal
For absolute none percentage values, the formatting should be a noop.
Changed the NoFormatter format method to not use the NumberFormat util bc it automatically
rounds the number.

Also, modified the percentage formatter to only multiply by 100 if the value entered is between 0 and 1, previously it was multiplying without checking so if a user entered "99", the resulting value would be "9,900%"