[HHQ-3099] - org.hyperic.hq.ui.action.resource.service.inventory.ViewServiceAction needs to handle PermissionException
Handling the exception and displaying a more descriptive error message rather than the previous unknown exception message.

Ideally, this should be changed to retrieve the parent platform details in a limited, read-only manner if a user doesn't have full permissions to view the object.  It would provide a better experience since you'd still be able to do what you need to do regarding the platform service (which has the primary focus in this view) while preventing you from getting more information about the parent platform than is needed to function.