[HHQ-3194] - Alert definition creation: threshold value is set to 1% instead of 100%
Rolling back this change as the expected behavior is unclear.  The fix allowed users to enter a decimal value >0 and <1, which is converted to a value >0% and <100%.  If the user enters a value =0 or >=1, that value would be divided by 100 and then converted to a percentage.

The problem is users may have already set the value to 1 expecting it to =100%, which arguably, is a bug bc the logic that converts a value of 1 to 100%, also converts a value of 25 to 2,500%. In any case, the user will have the 1=100% expectation and this change would potentially create confusion if included.