[HHQ-2895] - Deleting a resource with recent control action breaks Recent Control Action portlet
Now catching the AppdefEntityNotFoundException which is thrown when trying to retrieve a resource that has been deleted.  If this exception is thrown, we catch it and remove the control history object from the list as it is no longer valid.  Another process later cleans this all up so that deleted objects aren't included in the list to begin with, but there is some latency between the delete action and the data cleanup.