[HHQ-3017] - Agent version is incorrectly displayed at the HQ-Health page after the remote upgrade of agent from HQ
Updated implementation of AgentUpgrade_result to support passing version, build and bundle name values
Modified the AgentCommandsClient interface to return a Map object when calling the upgrade method
Modified classes implementing the AgentCommandsClient interface to return a Map object
Modified the AgentManagerEJBImpl to take a AgentUpgrade_result return object, extract the version and update the EAM_AGENT table with the new version.
Changed the groovy script back to its original implementation

PLEASE NOTE: This change does not fix pre 4.2 agents. 
If upgrading from a < 4.2 version of the agent, the version column will not be updated in the database.  However, upgrades from 4.2 and beyond will update this field.