[HHQ-1677] - EmailAction fails with missing template during server startup
Modified startup_classes.txt in .org and EE to execute the UIStartupListener before the EscalationStartupListener

Modified the EscalationStartupListener to implement the PluginsDeployedCallback interface to load the escalation manager after the plugins have been deployed.

The UIStartupListener also implements this interface to load the HQU plugins and the list of callbacks are executed in the order in which they are registered, hence the changes to start_classes.txt.